Add files via upload
This version of commit makes a list of nbr trained from trainer.py file and uses it to find a match with the nbr predicted from the detector.py file. 

It also includes the functionality of deleting the data from the dataSet once the trainer gets executed for the set and dumps the files to data-dump directory. This gives less execution time for the trainer module as it only goes through the dataSet still remaining to get trained.